Swap between playing & not playing.
