+++ src/kube-cards-website/src/components/play/KubeCardsPlayGame.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const KubeCardsPlayGame = () => {
+    return (
+        <div>
+            <h1>Now playing a game!</h1>
+        </div>
+    );
+};
+
+export default KubeCardsPlayGame;
